app/back/requirements.txt
db-dtypes==1.0.4
elasticsearch==8.5.1
fastapi==0.89.1
google-auth==2.17.3
google-cloud-bigquery==3.10.0
huggingface_hub==0.13.4
numpy==1.23.5
pandas==1.5.3
pyarrow==9.0.0
python-dotenv==1.0.0
sentence-transformers==2.2.2
uvicorn==0.20.0
loguru==0.6.0